requirements.txt
aiohttp
requests
ollama
python-dotenv
fastapi
uvicorn[standard]
gunicorn
slowapi